// types/generated/graphql.tsx
export type Maybe<T> = T | null
export type Exact<T extends { [key: string]: unknown }> = {
  [K in keyof T]: T[K]
}
export type MakeOptional<T, K extends keyof T> = Omit<T, K> &
  { [SubKey in K]?: Maybe<T[SubKey]> }
export type MakeMaybe<T, K extends keyof T> = Omit<T, K> &
  { [SubKey in K]: Maybe<T[SubKey]> }
const defaultOptions = {}
/** All built-in and custom scalars, mapped to their actual values */
export type Scalars = {
  ID: string
  String: string
  Boolean: boolean
  Int: number
  Float: number
  timestamptz: any
  uuid: string
}
// ...
export type Groups = {
  __typename?: 'groups'
  id: Scalars['uuid']
  name: Scalars['String']
  /** An array relationship */
  users: Array<Users>
  /** An aggregate relationship */
  users_aggregate: Users_Aggregate
}
// ...
export type Profile_Users = {
  __typename?: 'profile_users'
  id: Scalars['uuid']
  /** An object relationship */
  profile: Profiles
  profile_id: Scalars['uuid']
  /** An object relationship */
  user: Users
  user_id: Scalars['uuid']
}

/** aggregated selection of "profile_users" */
export type Profile_Users_Aggregate = {
  __typename?: 'profile_users_aggregate'
  aggregate?: Maybe<Profile_Users_Aggregate_Fields>
  nodes: Array<Profile_Users>
}
// ...
export type Profiles = {
  __typename?: 'profiles'
  id: Scalars['uuid']
  nickname: Scalars['String']
  /** fetch data from the table: "profile_users" */
  profile_users: Array<Profile_Users>
  /** An aggregate relationship */
  profile_users_aggregate: Profile_Users_Aggregate
  /** An array relationship */
  users: Array<Users>
  /** An aggregate relationship */
  users_aggregate: Users_Aggregate
}
// ...
export type Users = {
  __typename?: 'users'
  created_at: Scalars['timestamptz']
  /** An object relationship */
  group?: Maybe<Groups>
  group_id?: Maybe<Scalars['uuid']>
  id: Scalars['uuid']
  name: Scalars['String']
  /** An object relationship */
  profile?: Maybe<Profiles>
  profile_id?: Maybe<Scalars['uuid']>
  /** fetch data from the table: "profile_users" */
  profile_users: Array<Profile_Users>
  /** An aggregate relationship */
  profile_users_aggregate: Profile_Users_Aggregate
}
// ...
export type GetUsersQuery = { __typename?: 'query_root' } & {
  users: Array<
    { __typename?: 'users' } & Pick<Users, 'id' | 'name' | 'created_at'>
  >
}
// ...
export type CreateUserMutation = { __typename?: 'mutation_root' } & {
  insert_users_one?: Maybe<
    { __typename?: 'users' } & Pick<Users, 'id' | 'name' | 'created_at'>
  >
}

export type DeleteUserMutationVariables = Exact<{
  id: Scalars['uuid']
}>

export type DeleteUserMutation = { __typename?: 'mutation_root' } & {
  delete_users_by_pk?: Maybe<
    { __typename?: 'users' } & Pick<Users, 'id' | 'name' | 'created_at'>
  >
}

export type UpdateUserMutationVariables = Exact<{
  id: Scalars['uuid']
  name: Scalars['String']
}>

export type UpdateUserMutation = { __typename?: 'mutation_root' } & {
  update_users_by_pk?: Maybe<
    { __typename?: 'users' } & Pick<Users, 'id' | 'name' | 'created_at'>
  >
}